new encoding.cpp
