+++ inst/include/tools/encoding.h
@@ -1,0 +1,11 @@
+#ifndef dplyr_tools_encoding_H
+#define dplyr_tools_encoding_H
+
+namespace dplyr {
+
+CharacterVector reencode_factor(IntegerVector x);
+CharacterVector reencode_char(SEXP x);
+
+}
+
+#endif // #ifndef dplyr_tools_encoding_H
